+++ components/Simulator/SimulatorButtons.js
@@ -3,10 +3,10 @@ import { memo, useContext, useState } from "react";
 import { useTranslations } from "next-intl";
 import { FaCheck, FaRegCopy, FaArrowUpRightFromSquare } from "react-icons/fa6";
 import Button from "@/components/Button";
+import ConfirmModal from "@/components/ConfirmModal";
 import LoadoutContext from "@/contexts/LoadoutContext";
 import ModalContext from "@/contexts/ModalContext";
 import styles from "./Simulator.module.scss";
-import ConfirmModal from "../ConfirmModal";
 
 function SimulatorButtons() {
   const t = useTranslations("SimulatorButtons");