+++ simulator/engine/StageEngine.js
@@ -488,7 +488,6 @@ export default class StageEngine {
       }
       state.effects.push({ ...effect, sourceType, sourceId });
     }
-    state.effects.sort((a, b) => (a.order || 0) - (b.order || 0)); // Ascending
     return state;
   }
 
@@ -502,6 +501,7 @@ export default class StageEngine {
       if (effect.phase != phase) continue;
       phaseEffects.push({ ...effect, phase: null, index: i });
     }
+    phaseEffects.sort((a, b) => (a.order || 0) - (b.order || 0));
 
     this.logger.debug(phase, phaseEffects);