+++ simulator/strategies/index.js
@@ -1,8 +1,10 @@
 import HeuristicStrategy from "./HeuristicStrategy";
+import HeuristicStrategyS8 from "./HeuristicStrategyS8";
 import RandomStrategy from "./RandomStrategy";
 
 const STRATEGIES = {
   HeuristicStrategy,
+  HeuristicStrategyS8,
   RandomStrategy,
 };